Reorganize files and packages
