+++ theme-orange-country/src/main/java/com/orange/ouds/theme/orangecountry/tokens/raw/OrangeCountryRawColors.kt
@@ -1,0 +1,20 @@
+/*
+ * Software Name: OUDS Android
+ * SPDX-FileCopyrightText: Copyright (c) Orange SA
+ * SPDX-License-Identifier: MIT
+ *
+ * This software is distributed under the MIT license,
+ * the text of which is available at https://opensource.org/license/MIT/
+ * or see the "LICENSE" file for more details.
+ *
+ * Software description: Android library of reusable graphical components
+ */
+
+package com.orange.ouds.theme.orangecountry.tokens.raw
+
+import androidx.compose.ui.graphics.Color
+
+internal object OrangeCountryRawColors {
+    val alert100 = Color(0xffFFCC00)
+    val alert200 = Color(0xff8F7200)
+}
